refactor(auth): move signup mutation to hooks directory
The signup mutation logic was moved from the api mutations file to a dedicated hook file in the features directory to better organize the auth-related hooks and follow the project's hook structure.
